requirements.txt
# Code-RL-Ground Python Requirements
# ==================================

# Core ML/RL
torch>=2.1.0
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Data & Config
pyyaml>=6.0
numpy>=1.24.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.21.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
